requirements.txt
discord.py==2.5.2
pytz==2021.1
python-dotenv==1.1.1
legacy-cgi==2.6.2
audioop-lts==0.2.1